refactor: split complex pages into page/container/ui pattern
- mentor-pool: extract data fetching and state into container, UI into ui
- profile/[pageUserId]: extract hooks/handlers into container, JSX into ui
- profile/card: extract session + data fetching into container, JSX into ui
- auth/signup: rename AuthPage to Page for consistency
